# Limiter: avoid stop-the-world pause when reading heap size