#### What this PR does

Use a Go metric instead of calling `runtime.ReadMemStats`. Empirically this is much faster.

```
                  │ before.txt  │             after.txt              │
                  │   sec/op    │   sec/op     vs base               │
CombinedScanner-4   25.58µ ± 2%   12.41µ ± 3%  -51.48% (p=0.002 n=6)
```

In a trace from a large ingester; the stop-the-world pause was more like 1-5ms; see example in image.
(and this is called every 1 second)

![image](https://github.com/user-attachments/assets/bee81b23-476b-45fd-af19-58600a46a22d)


#### Checklist

- NA Tests updated.
- NA Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- NA [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.